+++ CHANGELOG.md
@@ -4,3 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
+
+## [Unreleased]
+
+### Added
+
+- Added `smee` command
+- Added `install` subcommand to `smee` command
+- Added `uninstall` subcommand to `smee` command
+- Added `hook` subcommand to `smee` command
+- Added `help` subcommand to `smee` command
+- Added `--version` option to `smee` command
+- Added `--help` option to `smee` command